rename nipoppy.tabular.doughnut module
